pyproject.toml
[project]
name = "neo4j-text2sql"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.118.2",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "neo4j>=6.0.2",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "sqlglot>=27.24.2",
    "uvicorn>=0.37.0",
]
